Make sure we only connect to SMTP on-demand
Summary:
When running the tests, we expect the local MTA to be working. That's not something that we can assume.

The notification tests all mock the SMTP object to check whether it's called
properly, so we don't really need to connect on backend initialization.

Test Plan:
Run tests with local MTA disabled, make sure they still pass.

Group Reviewers: Reviewers

Reviewers: ardumont

Reviewed By: ardumont

Differential Revision: https://forge.softwareheritage.org/D308