tests: make db-related tests use sql files from the package
this require a small change in vault's swh-schema.sql: create the dbversion
tqble only if needed, otherwise this sql cannot be executed without
error on the top of storages' db schema. I guess it used to work thanks to
the fact theses used to be pg_restored instead of SQL statements executed via psql (with '-v ON_ERROR_STOP=1', which se do want).

Related to T1265.