rpc/server: compute make_app_from_configfile's default value at exec time
for the config_file argument, rather than declaration time.

This is required so that if this function is called with None as first
argument (which is what happens in cli.py when user do not specify a
config file option).

Long story short, this fixes the rpc-server cli command: if a config file
is not explicitely given as option, fall back to default config file search
behavior instead of crashing.