to_disk: raise an AssertionError if content has an unexpected status.
Although that will be fixed in the next version of swh-storage,
the current one returns a None status for skipped content, which
causes surprising behavior from the cooker (as it reuse the content
of the previous file in place of the current one).

Additionally, even though it shouldn't happen, if a content is
missing completely (ie. neither know nor skipped), it's still
possible to have a None status.