Rename obj_type to bundle_type
'obj_type' was confusing, because it means 'the type of objects in the
vault backend', which is actually made of an object type (dir/rev/...)
AND the cooker name.

Plus, I want to reuse the name 'obj_type' in a future commit to reference
what is actually the object type, in the swh-model definition.