git_bare: Send progress updates while cooking
It can be confusing to just have 'Processing...' in the UI for hours.
With this change, a new progress update should be sent every few minutes.

In the swh-web UI, it will show only 'Processing... %d objects processed'
in the main table, but 'Over %d remaining' will be shown when hovering
the progress bar.
This does not show 'Processing... %d/%d' or a percentage, because the
number of remaining objects is not always known.