backend: Insert cooking task info in database after task creation
It might happen that the scheduler RPC service is failing when attempting
to submit a vault cooking task.

So prefer to insert cooking task info in vault database once the task
successfully created by the scheduler. Previously, the non created
cooking task was inserted in database with a null task id and a new
status which prevented to create a new one once the scheduler service
rpc issue fixed.

Related to swh/infra/sysadm-environment#5055.