cookers/git_bare: Speedup repository cooking with multi-threading
Previously when cooking a git bare repository, contents bytes were
fetched sequentially which could take a good amount of time for an
origin with a large revisions history.

In order to speedup the cooking process, retrieve the contents bytes
in parallel with the help of the concurrent.futures module from the
Python standard library which fits particularly well for making loops
of I/O-bound tasks concurrent and for issuing tasks asynchronously.