backend: Force bundle recooking if it is no longer in cache
It can exist cases when a bundle has been cooked and marked as done in
the vault database but is no longer available in the vault cache.

Previously, when requesting a new cooking of such bundle, the vault
backend was not creating a new cooking task and thus the bundle could
not be downloaded anymore.

So ensure to check bundle presence in cache for such edge case and force
its recooking if it is not the case.

Related to swh/devel/swh-web#4804.