vault: Relax backend instantiation configuration constraints
For the migration context case, it's ok to relax the constraints on the
configuration not specifying the cache, scheduler or storage key entries.

It's not the case for the actual runtime case. So assertions in the code using
those backends has been added to fail as soon as possible.

Refs. swh/infra/sysadm-environment#5562